backend/requirements.txt
# === Core Framework ===
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.28.1

# === Database ===
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.1
motor==3.6.0        # Async MongoDB driver

# === Auth ===
pyjwt==2.10.1
bcrypt==4.2.1
cryptography==44.0.0
eth-account<0.13.0   # Web3 signature verification
web3==6.20.1         # Main Web3 library

# === Redis & Celery ===
redis==5.2.1
celery[redis]==5.4.0

# === AI/ML ===
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.6.1
tensorflow==2.18.0
yfinance==0.2.51

# === NLP ===
transformers==4.47.1
torch==2.5.1

# === Gemini AI ===
google-generativeai==0.8.4

# === News ===
newsapi-python==0.2.7

# === Monitoring ===
sentry-sdk[fastapi]==2.19.2
prometheus-fastapi-instrumentator==7.0.2

# === WebSocket ===
websockets==14.1

# === Dev ===
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.10.0
flake8==7.1.1
